@@ -22,59 +22,57 @@ public static class ApiConfig
 
     public static readonly ReturnCodeType ReturnCodes = new()
     {
-        {0, (true, false, true, "Results successfully retrieved.")},
-        {1, (false, false, true, "No results.") },
-        {2, (true, false, true, "No match was found, displaying results for: ***{0}***.") },
-        {-1, (false, true, false, "{0}: Internal error occurred, please contact Ani and Nicba1010") },
-        {-2, (false, true, false, "{0}: API is undergoing maintenance, please try again later.") },
-        {-3, (false, false, false, "Illegal characters found, please try again with a different search term.") },
+        [0] = (true, false, true, "Results successfully retrieved."),
+        [1] = (false, false, true, "No results."),
+        [2] = (true, false, true, "No match was found, displaying results for: ***{0}***."),
+        [-1] = (false, true, false, "{0}: Internal error occurred, please contact Ani and Nicba1010"),
+        [-2] = (false, true, false, "{0}: API is undergoing maintenance, please try again later."),
+        [-3] = (false, false, false, "Illegal characters found, please try again with a different search term."),
     };
 
     public static readonly List<int> ResultAmount = [25, 50, 100, 200];
 
     public static readonly Dictionary<char, string[]> Directions = new()
     {
-        {'a', ["a", "asc", "ascending"] },
-        {'d', ["d", "desc", "descending"] },
+        ['a'] = ["a", "asc", "ascending"],
+        ['d'] = ["d", "desc", "descending"],
     };
 
     public static readonly Dictionary<string, int> Statuses = new()
     {
-        {"all", 0 },
-        {"playable", 1 },
-        {"ingame", 2 },
-        {"intro", 3 },
-        {"loadable", 4 },
-        {"nothing", 5 },
+        ["all"] = 0,
+        ["playable"] = 1,
+        ["ingame"] = 2,
+        ["intro"] = 3,
+        ["loadable"] = 4,
+        ["nothing"] = 5,
     };
 
     public static readonly Dictionary<string, int> SortTypes = new()
     {
-        {"id", 1 },
-        {"title", 2 },
-        {"status", 3 },
-        {"date", 4 },
+        ["id"] = 1,
+        ["title"] = 2,
+        ["status"] = 3,
+        ["date"] = 4,
     };
 
     public static readonly Dictionary<char, string[]> ReleaseTypes = new()
     {
-        {'b', ["b", "d", "disc", "disk", "bluray", "blu-ray"] },
-        {'n', ["n", "p", "PSN"] },
+        ['b'] = ["b", "d", "disc", "disk", "bluray", "blu-ray"],
+        ['n'] = ["n", "p", "PSN"],
     };
 
     public static readonly Dictionary<string, char> ReverseDirections;
     public static readonly Dictionary<string, char> ReverseReleaseTypes;
 
     private static Dictionary<TV, TK> Reverse<TK, TV>(this Dictionary<TK, TV[]> dic, IEqualityComparer<TV> comparer)
-        where TK: notnull
-        where TV: notnull
-    {
-        return (
+        where TK : notnull
+        where TV : notnull
+        => (
             from kvp in dic
             from val in kvp.Value
             select (val, kvp.Key)
         ).ToDictionary(rkvp => rkvp.val, rkvp => rkvp.Key, comparer);
-    }
 
     public static readonly ILogger Log;
     public static readonly RecyclableMemoryStreamManager MemoryStreamManager = new();
